colocate location module
