+++ apps/member-profile/app/shared/queries/index.ts
@@ -1,13 +1,10 @@
 import dayjs from 'dayjs';
 
+import { Country } from '@oyster/core/location.types';
 import { db } from '@oyster/db';
 import { toTitleCase } from '@oyster/utils';
 
-import {
-  Country,
-  type DegreeType,
-  FORMATTED_DEGREEE_TYPE,
-} from '@/member-profile.ui';
+import { type DegreeType, FORMATTED_DEGREEE_TYPE } from '@/member-profile.ui';
 
 // "educations"
