+++ packages/core/src/member-profile.server.ts
@@ -33,9 +33,6 @@ export { listActivities } from './modules/gamification/queries/list-activities';
 export { getIcebreakerPrompts } from './modules/icebreaker/queries/get-icebreaker-prompts';
 export { getIcebreakerResponses } from './modules/icebreaker/queries/get-icebreaker-responses';
 export { upsertIcebreakerResponses } from './modules/icebreaker/use-cases/upsert-icebreaker-responses';
-export { getAutocompletedCities } from './modules/location/queries/get-autocompleted-cities';
-export { getCityDetails } from './modules/location/queries/get-city-details';
-export { searchCountries } from './modules/location/queries/search-countries';
 export { listEmails } from './modules/member/queries/list-emails';
 export { listMembersInDirectory } from './modules/member/queries/list-members-in-directory';
 export { changePrimaryEmail } from './modules/member/use-cases/change-primary-email';
